requirements.txt
openai
gradio
python-dotenv
llama_index
pinecone